fixed unit test for uploadPackage
